+++ src/photon_platform/publish/app.py
@@ -22,3 +22,4 @@ def test(port):
     publish.test(port=port)
 
 
+